requirements.txt
opencv-python
numpy
python-docx
Pillow